requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.1
h11==0.14.0
httptools==0.5.0
idna==3.4
psycopg2-binary==2.9.6
pyaes==1.6.1
pyasn1==0.5.0
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
rsa==4.9
sniffio==1.3.0
SQLAlchemy==2.0.11
starlette==0.26.1
Telethon==1.28.5
typing_extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
